AI_News_Summary/requirements.txt
openai>=1.0
requests
python-dotenv
beautifulsoup4
